Updates & Fine Tuning
