+++ package/SKSE/Plugins/CommunityShaders/Themes/Amber.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Warm Amber",
-	"Description": "Cozy amber tones reminiscent of hearth fires and candlelight",
-	"Version": "1.0.0",
+	"Description": "Cozy amber tones reminiscent of hearth fires and candlelight in Nordic taverns",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 26.8,
+		"GlobalScale": -0.02,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.5,
 		"Palette": {
 			"Background": [0.2, 0.15, 0.05, 0.9],
 			"Text": [1.0, 0.9, 0.7, 1.0],
 			"Border": [0.8, 0.6, 0.3, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.5, 0.4, 0.3, 1.0],
+			"Error": [1.0, 0.4, 0.2, 1.0],
+			"Warning": [1.0, 0.7, 0.1, 1.0],
+			"RestartNeeded": [0.8, 0.9, 0.3, 1.0],
+			"CurrentHotkey": [1.0, 0.8, 0.4, 1.0],
+			"SuccessColor": [0.6, 0.8, 0.3, 1.0],
+			"InfoColor": [0.7, 0.8, 0.9, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.9, 0.8, 0.6, 1.0],
+			"ColorHovered": [1.0, 0.9, 0.7, 1.0],
+			"MinimizedFactor": 0.65
+		},
+		"Style": {
+			"WindowBorderSize": 2.0,
+			"ChildBorderSize": 1.0,
+			"FrameBorderSize": 1.5,
+			"WindowPadding": [14.0, 12.0],
+			"WindowRounding": 5.0,
+			"IndentSpacing": 8.0,
+			"FramePadding": [7.0, 5.0],
+			"CellPadding": [12.0, 5.0],
+			"ItemSpacing": [8.0, 8.0]
+		},
+		"FullPalette": [
+			[1.00, 0.90, 0.70, 0.90],
+			[0.80, 0.70, 0.50, 1.00],
+			[0.20, 0.15, 0.05, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.18, 0.13, 0.04, 0.85],
+			[0.70, 0.60, 0.40, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.15, 0.10, 0.03, 1.00],
+			[0.40, 0.30, 0.15, 0.40],
+			[0.50, 0.40, 0.20, 0.45],
+			[0.15, 0.10, 0.03, 0.83],
+			[0.18, 0.13, 0.05, 0.87],
+			[0.30, 0.22, 0.10, 0.90],
+			[0.22, 0.16, 0.06, 0.90],
+			[0.35, 0.26, 0.12, 0.90],
+			[0.45, 0.35, 0.18, 1.00],
+			[0.55, 0.45, 0.25, 1.00],
+			[0.65, 0.55, 0.35, 1.00],
+			[1.00, 0.90, 0.70, 1.00],
+			[0.80, 0.70, 0.50, 1.00],
+			[0.40, 0.30, 0.15, 1.00],
+			[0.80, 0.60, 0.30, 0.40],
+			[0.90, 0.70, 0.40, 0.67],
+			[1.00, 0.80, 0.50, 1.00],
+			[0.95, 0.65, 0.25, 1.00],
+			[1.00, 0.75, 0.35, 1.00],
+			[1.00, 0.85, 0.45, 1.00],
+			[0.70, 0.55, 0.30, 1.00],
+			[0.80, 0.65, 0.40, 1.00],
+			[0.95, 0.80, 0.55, 1.00],
+			[1.00, 0.90, 0.70, 1.00],
+			[1.00, 0.90, 0.70, 0.60],
+			[1.00, 0.90, 0.70, 0.90],
+			[0.80, 0.60, 0.30, 0.31],
+			[0.90, 0.70, 0.40, 0.80],
+			[1.00, 0.80, 0.50, 1.00],
+			[0.18, 0.13, 0.04, 0.97],
+			[0.85, 0.70, 0.40, 1.00],
+			[0.70, 0.55, 0.30, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 0.90, 0.70, 1.00],
+			[1.00, 0.80, 0.30, 1.00],
+			[1.00, 0.80, 0.30, 1.00],
+			[1.00, 0.80, 0.30, 1.00],
+			[0.80, 0.60, 0.30, 0.40],
+			[0.35, 0.26, 0.12, 1.00],
+			[0.28, 0.20, 0.09, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 0.90, 0.70, 0.06],
+			[0.80, 0.60, 0.30, 0.35],
+			[0.90, 0.50, 0.30, 1.00],
+			[0.90, 0.70, 0.40, 1.00],
+			[0.50, 0.40, 0.20, 0.56],
+			[0.35, 0.26, 0.12, 0.35],
+			[0.35, 0.26, 0.12, 0.35]
+		]
 	}
 }
