+++ package/SKSE/Plugins/CommunityShaders/Themes/Default.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Default Dark",
-	"Description": "The classic Community Shaders dark theme",
-	"Version": "1.0.0",
+	"Description": "The classic Community Shaders dark theme with comprehensive styling",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 27.0,
+		"GlobalScale": 0.0,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.5,
 		"Palette": {
 			"Background": [0.09, 0.09, 0.09, 0.95],
 			"Text": [1.0, 1.0, 1.0, 1.0],
 			"Border": [0.5, 0.5, 0.5, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.5, 0.5, 0.5, 1.0],
+			"Error": [1.0, 0.4, 0.4, 1.0],
+			"Warning": [1.0, 0.6, 0.2, 1.0],
+			"RestartNeeded": [0.4, 1.0, 0.4, 1.0],
+			"CurrentHotkey": [1.0, 1.0, 0.0, 1.0],
+			"SuccessColor": [0.0, 1.0, 0.0, 1.0],
+			"InfoColor": [0.2, 0.6, 1.0, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.8, 0.8, 0.8, 1.0],
+			"ColorHovered": [0.6, 0.6, 0.6, 1.0],
+			"MinimizedFactor": 0.7
+		},
+		"Style": {
+			"WindowBorderSize": 2.0,
+			"ChildBorderSize": 0.0,
+			"FrameBorderSize": 1.0,
+			"WindowPadding": [12.0, 12.0],
+			"WindowRounding": 4.0,
+			"IndentSpacing": 8.0,
+			"FramePadding": [6.0, 4.0],
+			"CellPadding": [12.0, 4.0],
+			"ItemSpacing": [8.0, 8.0]
+		},
+		"FullPalette": [
+			[0.90, 0.90, 0.90, 0.90],
+			[0.60, 0.60, 0.60, 1.00],
+			[0.09, 0.09, 0.09, 0.95],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.05, 0.05, 0.10, 0.85],
+			[0.70, 0.70, 0.70, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.00, 0.00, 0.00, 1.00],
+			[0.26, 0.26, 0.26, 0.40],
+			[0.40, 0.40, 0.40, 0.45],
+			[0.00, 0.00, 0.00, 0.83],
+			[0.00, 0.00, 0.00, 0.87],
+			[0.20, 0.20, 0.30, 0.90],
+			[0.02, 0.02, 0.03, 0.90],
+			[0.20, 0.22, 0.27, 0.90],
+			[0.28, 0.28, 0.28, 1.00],
+			[0.42, 0.42, 0.42, 1.00],
+			[0.56, 0.56, 0.56, 1.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.70, 0.70, 0.70, 1.00],
+			[0.26, 0.26, 0.26, 1.00],
+			[0.26, 0.59, 0.98, 0.40],
+			[0.26, 0.59, 0.98, 0.67],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.06, 0.53, 0.98, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.50, 0.50, 0.50, 1.00],
+			[0.70, 0.60, 0.60, 1.00],
+			[0.90, 0.70, 0.70, 1.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[1.00, 1.00, 1.00, 0.60],
+			[1.00, 1.00, 1.00, 0.90],
+			[0.26, 0.59, 0.98, 0.31],
+			[0.26, 0.59, 0.98, 0.80],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.15, 0.15, 0.15, 0.97],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.70, 0.60, 0.60, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.26, 0.59, 0.98, 0.40],
+			[0.26, 0.26, 0.26, 1.00],
+			[0.19, 0.19, 0.19, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 1.00, 1.00, 0.06],
+			[0.26, 0.59, 0.98, 0.35],
+			[0.80, 0.50, 0.50, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.30, 0.30, 0.30, 0.56],
+			[0.20, 0.20, 0.20, 0.35],
+			[0.20, 0.20, 0.20, 0.35]
+		]
 	}
 }
