+++ package/SKSE/Plugins/CommunityShaders/Themes/DragonBlood.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Dragon Blood",
 	"Description": "Dark red theme inspired by dragon lore and ancient power",
-	"Version": "1.0.0",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 28.0,
+		"GlobalScale": 0.1,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.4,
 		"Palette": {
 			"Background": [0.25, 0.05, 0.05, 0.9],
 			"Text": [1.0, 0.85, 0.85, 1.0],
 			"Border": [0.8, 0.3, 0.3, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.4, 0.2, 0.2, 1.0],
+			"Error": [1.0, 0.1, 0.1, 1.0],
+			"Warning": [1.0, 0.5, 0.0, 1.0],
+			"RestartNeeded": [0.8, 0.6, 0.2, 1.0],
+			"CurrentHotkey": [1.0, 0.8, 0.2, 1.0],
+			"SuccessColor": [0.6, 0.8, 0.2, 1.0],
+			"InfoColor": [0.8, 0.4, 0.6, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.9, 0.6, 0.6, 1.0],
+			"ColorHovered": [1.0, 0.7, 0.7, 1.0],
+			"MinimizedFactor": 0.6
+		},
+		"Style": {
+			"WindowBorderSize": 3.0,
+			"ChildBorderSize": 1.0,
+			"FrameBorderSize": 2.0,
+			"WindowPadding": [16.0, 14.0],
+			"WindowRounding": 8.0,
+			"IndentSpacing": 10.0,
+			"FramePadding": [8.0, 6.0],
+			"CellPadding": [14.0, 6.0],
+			"ItemSpacing": [10.0, 10.0]
+		},
+		"FullPalette": [
+			[1.00, 0.85, 0.85, 0.90],
+			[0.80, 0.60, 0.60, 1.00],
+			[0.25, 0.05, 0.05, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.20, 0.03, 0.03, 0.85],
+			[0.60, 0.40, 0.40, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.15, 0.00, 0.00, 1.00],
+			[0.40, 0.15, 0.15, 0.40],
+			[0.50, 0.20, 0.20, 0.45],
+			[0.10, 0.00, 0.00, 0.83],
+			[0.15, 0.00, 0.00, 0.87],
+			[0.30, 0.10, 0.10, 0.90],
+			[0.25, 0.05, 0.05, 0.90],
+			[0.35, 0.15, 0.15, 0.90],
+			[0.45, 0.15, 0.15, 1.00],
+			[0.60, 0.25, 0.25, 1.00],
+			[0.75, 0.35, 0.35, 1.00],
+			[1.00, 0.85, 0.85, 1.00],
+			[0.80, 0.60, 0.60, 1.00],
+			[0.40, 0.15, 0.15, 1.00],
+			[0.80, 0.30, 0.30, 0.40],
+			[0.90, 0.40, 0.40, 0.67],
+			[1.00, 0.50, 0.50, 1.00],
+			[0.85, 0.25, 0.25, 1.00],
+			[0.90, 0.35, 0.35, 1.00],
+			[0.95, 0.45, 0.45, 1.00],
+			[0.60, 0.30, 0.30, 1.00],
+			[0.80, 0.40, 0.40, 1.00],
+			[0.95, 0.55, 0.55, 1.00],
+			[1.00, 0.85, 0.85, 1.00],
+			[1.00, 0.85, 0.85, 0.60],
+			[1.00, 0.85, 0.85, 0.90],
+			[0.80, 0.30, 0.30, 0.31],
+			[0.90, 0.40, 0.40, 0.80],
+			[1.00, 0.50, 0.50, 1.00],
+			[0.20, 0.05, 0.05, 0.97],
+			[0.85, 0.35, 0.35, 1.00],
+			[0.70, 0.30, 0.30, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 0.85, 0.85, 1.00],
+			[1.00, 0.60, 0.20, 1.00],
+			[1.00, 0.60, 0.20, 1.00],
+			[1.00, 0.60, 0.20, 1.00],
+			[0.80, 0.30, 0.30, 0.40],
+			[0.40, 0.15, 0.15, 1.00],
+			[0.30, 0.10, 0.10, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 0.85, 0.85, 0.06],
+			[0.80, 0.30, 0.30, 0.35],
+			[1.00, 0.30, 0.30, 1.00],
+			[0.90, 0.40, 0.40, 1.00],
+			[0.50, 0.20, 0.20, 0.56],
+			[0.35, 0.15, 0.15, 0.35],
+			[0.35, 0.15, 0.15, 0.35]
+		]
 	}
 }
