+++ package/SKSE/Plugins/CommunityShaders/Themes/DwemerBronze.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Dwemer Bronze",
-	"Description": "Ancient bronze theme inspired by lost Dwemer technology",
-	"Version": "1.0.0",
+	"Description": "Ancient bronze theme inspired by lost Dwemer technology and metallic machinery",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 27.5,
+		"GlobalScale": 0.05,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.3,
 		"Palette": {
 			"Background": [0.15, 0.12, 0.08, 0.9],
 			"Text": [0.9, 0.75, 0.5, 1.0],
 			"Border": [0.7, 0.5, 0.3, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.4, 0.35, 0.25, 1.0],
+			"Error": [0.9, 0.3, 0.1, 1.0],
+			"Warning": [1.0, 0.7, 0.2, 1.0],
+			"RestartNeeded": [0.8, 0.8, 0.4, 1.0],
+			"CurrentHotkey": [1.0, 0.8, 0.3, 1.0],
+			"SuccessColor": [0.5, 0.7, 0.3, 1.0],
+			"InfoColor": [0.6, 0.7, 0.8, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.8, 0.65, 0.4, 1.0],
+			"ColorHovered": [0.95, 0.8, 0.55, 1.0],
+			"MinimizedFactor": 0.7
+		},
+		"Style": {
+			"WindowBorderSize": 3.0,
+			"ChildBorderSize": 2.0,
+			"FrameBorderSize": 2.0,
+			"WindowPadding": [16.0, 14.0],
+			"WindowRounding": 2.0,
+			"IndentSpacing": 10.0,
+			"FramePadding": [8.0, 6.0],
+			"CellPadding": [14.0, 6.0],
+			"ItemSpacing": [10.0, 9.0]
+		},
+		"FullPalette": [
+			[0.90, 0.75, 0.50, 0.90],
+			[0.70, 0.60, 0.40, 1.00],
+			[0.15, 0.12, 0.08, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.13, 0.10, 0.06, 0.85],
+			[0.60, 0.50, 0.35, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.10, 0.08, 0.05, 1.00],
+			[0.35, 0.28, 0.18, 0.40],
+			[0.45, 0.36, 0.23, 0.45],
+			[0.10, 0.08, 0.05, 0.83],
+			[0.13, 0.10, 0.07, 0.87],
+			[0.25, 0.20, 0.13, 0.90],
+			[0.18, 0.14, 0.09, 0.90],
+			[0.30, 0.24, 0.15, 0.90],
+			[0.40, 0.32, 0.20, 1.00],
+			[0.50, 0.42, 0.28, 1.00],
+			[0.60, 0.52, 0.38, 1.00],
+			[0.90, 0.75, 0.50, 1.00],
+			[0.70, 0.60, 0.40, 1.00],
+			[0.35, 0.28, 0.18, 1.00],
+			[0.70, 0.50, 0.30, 0.40],
+			[0.80, 0.60, 0.35, 0.67],
+			[0.90, 0.70, 0.45, 1.00],
+			[0.85, 0.55, 0.25, 1.00],
+			[0.90, 0.65, 0.35, 1.00],
+			[0.95, 0.75, 0.45, 1.00],
+			[0.65, 0.50, 0.30, 1.00],
+			[0.75, 0.60, 0.40, 1.00],
+			[0.85, 0.70, 0.50, 1.00],
+			[0.90, 0.75, 0.50, 1.00],
+			[0.90, 0.75, 0.50, 0.60],
+			[0.90, 0.75, 0.50, 0.90],
+			[0.70, 0.50, 0.30, 0.31],
+			[0.80, 0.60, 0.35, 0.80],
+			[0.90, 0.70, 0.45, 1.00],
+			[0.13, 0.10, 0.06, 0.97],
+			[0.75, 0.60, 0.35, 1.00],
+			[0.65, 0.50, 0.30, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.75, 0.50, 1.00],
+			[0.90, 0.70, 0.30, 1.00],
+			[0.90, 0.70, 0.30, 1.00],
+			[0.90, 0.70, 0.30, 1.00],
+			[0.70, 0.50, 0.30, 0.40],
+			[0.30, 0.24, 0.15, 1.00],
+			[0.23, 0.18, 0.11, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.75, 0.50, 0.06],
+			[0.70, 0.50, 0.30, 0.35],
+			[0.80, 0.45, 0.25, 1.00],
+			[0.80, 0.60, 0.35, 1.00],
+			[0.45, 0.36, 0.23, 0.56],
+			[0.30, 0.24, 0.15, 0.35],
+			[0.30, 0.24, 0.15, 0.35]
+		]
 	}
 }
