+++ package/SKSE/Plugins/CommunityShaders/Themes/Forest.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Forest Green",
-	"Description": "Natural green theme inspired by Skyrim's ancient forests",
-	"Version": "1.0.0",
+	"Description": "Natural green theme inspired by Skyrim's ancient forests and wilderness",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 26.5,
+		"GlobalScale": -0.05,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.6,
 		"Palette": {
 			"Background": [0.1, 0.3, 0.15, 0.9],
 			"Text": [0.9, 1.0, 0.9, 1.0],
 			"Border": [0.4, 0.7, 0.4, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.3, 0.4, 0.3, 1.0],
+			"Error": [0.8, 0.3, 0.2, 1.0],
+			"Warning": [0.9, 0.7, 0.2, 1.0],
+			"RestartNeeded": [0.6, 0.9, 0.3, 1.0],
+			"CurrentHotkey": [0.8, 1.0, 0.6, 1.0],
+			"SuccessColor": [0.2, 0.8, 0.3, 1.0],
+			"InfoColor": [0.4, 0.8, 0.9, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.7, 0.9, 0.7, 1.0],
+			"ColorHovered": [0.8, 1.0, 0.8, 1.0],
+			"MinimizedFactor": 0.65
+		},
+		"Style": {
+			"WindowBorderSize": 2.5,
+			"ChildBorderSize": 1.0,
+			"FrameBorderSize": 1.5,
+			"WindowPadding": [14.0, 12.0],
+			"WindowRounding": 6.0,
+			"IndentSpacing": 9.0,
+			"FramePadding": [7.0, 5.0],
+			"CellPadding": [13.0, 5.0],
+			"ItemSpacing": [9.0, 9.0]
+		},
+		"FullPalette": [
+			[0.90, 1.00, 0.90, 0.90],
+			[0.60, 0.80, 0.60, 1.00],
+			[0.10, 0.30, 0.15, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.08, 0.25, 0.12, 0.85],
+			[0.50, 0.70, 0.50, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.05, 0.15, 0.05, 1.00],
+			[0.20, 0.40, 0.25, 0.40],
+			[0.30, 0.50, 0.35, 0.45],
+			[0.05, 0.15, 0.08, 0.83],
+			[0.08, 0.20, 0.10, 0.87],
+			[0.15, 0.35, 0.20, 0.90],
+			[0.10, 0.25, 0.15, 0.90],
+			[0.18, 0.38, 0.22, 0.90],
+			[0.25, 0.45, 0.30, 1.00],
+			[0.35, 0.55, 0.40, 1.00],
+			[0.45, 0.65, 0.50, 1.00],
+			[0.90, 1.00, 0.90, 1.00],
+			[0.70, 0.85, 0.75, 1.00],
+			[0.20, 0.40, 0.25, 1.00],
+			[0.40, 0.70, 0.45, 0.40],
+			[0.50, 0.80, 0.55, 0.67],
+			[0.60, 0.90, 0.65, 1.00],
+			[0.35, 0.75, 0.40, 1.00],
+			[0.45, 0.85, 0.50, 1.00],
+			[0.55, 0.95, 0.60, 1.00],
+			[0.40, 0.60, 0.45, 1.00],
+			[0.50, 0.70, 0.55, 1.00],
+			[0.70, 0.90, 0.75, 1.00],
+			[0.90, 1.00, 0.90, 1.00],
+			[0.90, 1.00, 0.90, 0.60],
+			[0.90, 1.00, 0.90, 0.90],
+			[0.40, 0.70, 0.45, 0.31],
+			[0.50, 0.80, 0.55, 0.80],
+			[0.60, 0.90, 0.65, 1.00],
+			[0.12, 0.22, 0.15, 0.97],
+			[0.45, 0.75, 0.50, 1.00],
+			[0.50, 0.70, 0.55, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 1.00, 0.90, 1.00],
+			[0.80, 0.90, 0.20, 1.00],
+			[0.80, 0.90, 0.20, 1.00],
+			[0.80, 0.90, 0.20, 1.00],
+			[0.40, 0.70, 0.45, 0.40],
+			[0.20, 0.35, 0.25, 1.00],
+			[0.15, 0.28, 0.20, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 1.00, 0.90, 0.06],
+			[0.40, 0.70, 0.45, 0.35],
+			[0.70, 0.40, 0.30, 1.00],
+			[0.50, 0.80, 0.55, 1.00],
+			[0.25, 0.45, 0.30, 0.56],
+			[0.18, 0.35, 0.22, 0.35],
+			[0.18, 0.35, 0.22, 0.35]
+		]
 	}
 }
