+++ package/SKSE/Plugins/CommunityShaders/Themes/HighContrast.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "High Contrast",
-	"Description": "High contrast theme for improved accessibility and visibility",
-	"Version": "1.0.0",
+	"Description": "High contrast black and white theme for improved accessibility and visibility",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 29.0,
+		"GlobalScale": 0.05,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.3,
 		"Palette": {
 			"Background": [0.0, 0.0, 0.0, 0.95],
 			"Text": [1.0, 1.0, 1.0, 1.0],
 			"Border": [1.0, 1.0, 1.0, 0.9]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.5, 0.5, 0.5, 1.0],
+			"Error": [1.0, 0.0, 0.0, 1.0],
+			"Warning": [1.0, 1.0, 0.0, 1.0],
+			"RestartNeeded": [0.0, 1.0, 0.0, 1.0],
+			"CurrentHotkey": [0.0, 1.0, 1.0, 1.0],
+			"SuccessColor": [0.0, 1.0, 0.0, 1.0],
+			"InfoColor": [0.0, 0.5, 1.0, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [1.0, 1.0, 1.0, 1.0],
+			"ColorHovered": [0.8, 0.8, 0.8, 1.0],
+			"MinimizedFactor": 0.6
+		},
+		"Style": {
+			"WindowBorderSize": 4.0,
+			"ChildBorderSize": 2.0,
+			"FrameBorderSize": 2.0,
+			"WindowPadding": [18.0, 16.0],
+			"WindowRounding": 0.0,
+			"IndentSpacing": 12.0,
+			"FramePadding": [10.0, 6.0],
+			"CellPadding": [16.0, 8.0],
+			"ItemSpacing": [12.0, 12.0]
+		},
+		"FullPalette": [
+			[1.00, 1.00, 1.00, 0.90],
+			[0.80, 0.80, 0.80, 1.00],
+			[0.00, 0.00, 0.00, 0.95],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.05, 0.05, 0.05, 0.85],
+			[1.00, 1.00, 1.00, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.00, 0.00, 0.00, 1.00],
+			[0.30, 0.30, 0.30, 0.40],
+			[0.50, 0.50, 0.50, 0.45],
+			[0.00, 0.00, 0.00, 0.83],
+			[0.00, 0.00, 0.00, 0.87],
+			[0.20, 0.20, 0.20, 0.90],
+			[0.00, 0.00, 0.00, 0.90],
+			[0.15, 0.15, 0.15, 0.90],
+			[0.30, 0.30, 0.30, 1.00],
+			[0.50, 0.50, 0.50, 1.00],
+			[0.70, 0.70, 0.70, 1.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.80, 0.80, 0.80, 1.00],
+			[0.20, 0.20, 0.20, 1.00],
+			[0.60, 0.60, 0.60, 0.40],
+			[0.80, 0.80, 0.80, 0.67],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.90, 0.90, 0.90, 1.00],
+			[0.85, 0.85, 0.85, 1.00],
+			[0.75, 0.75, 0.75, 1.00],
+			[0.60, 0.60, 0.60, 1.00],
+			[0.80, 0.80, 0.80, 1.00],
+			[0.90, 0.90, 0.90, 1.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[1.00, 1.00, 1.00, 0.60],
+			[1.00, 1.00, 1.00, 0.90],
+			[0.60, 0.60, 0.60, 0.31],
+			[0.80, 0.80, 0.80, 0.80],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.10, 0.10, 0.10, 0.97],
+			[0.85, 0.85, 0.85, 1.00],
+			[0.70, 0.70, 0.70, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[1.00, 1.00, 0.00, 1.00],
+			[1.00, 1.00, 0.00, 1.00],
+			[1.00, 1.00, 0.00, 1.00],
+			[0.60, 0.60, 0.60, 0.40],
+			[0.20, 0.20, 0.20, 1.00],
+			[0.15, 0.15, 0.15, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[1.00, 1.00, 1.00, 0.06],
+			[0.60, 0.60, 0.60, 0.35],
+			[1.00, 0.00, 0.00, 1.00],
+			[0.80, 0.80, 0.80, 1.00],
+			[0.40, 0.40, 0.40, 0.56],
+			[0.25, 0.25, 0.25, 0.35],
+			[0.25, 0.25, 0.25, 0.35]
+		]
 	}
 }
