+++ package/SKSE/Plugins/CommunityShaders/Themes/Light.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Light Mode",
-	"Description": "Clean light theme with dark text for daytime use",
-	"Version": "1.0.0",
+	"Description": "Clean bright theme with dark text for comfortable daytime use",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 26.0,
+		"GlobalScale": -0.1,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.4,
 		"Palette": {
 			"Background": [0.9, 0.9, 0.9, 0.95],
 			"Text": [0.1, 0.1, 0.1, 1.0],
 			"Border": [0.3, 0.3, 0.3, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.6, 0.6, 0.6, 1.0],
+			"Error": [0.8, 0.2, 0.2, 1.0],
+			"Warning": [0.9, 0.5, 0.1, 1.0],
+			"RestartNeeded": [0.2, 0.7, 0.2, 1.0],
+			"CurrentHotkey": [0.8, 0.6, 0.1, 1.0],
+			"SuccessColor": [0.1, 0.6, 0.1, 1.0],
+			"InfoColor": [0.2, 0.4, 0.8, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.3, 0.3, 0.3, 1.0],
+			"ColorHovered": [0.2, 0.2, 0.2, 1.0],
+			"MinimizedFactor": 0.7
+		},
+		"Style": {
+			"WindowBorderSize": 1.5,
+			"ChildBorderSize": 0.5,
+			"FrameBorderSize": 1.0,
+			"WindowPadding": [12.0, 10.0],
+			"WindowRounding": 3.0,
+			"IndentSpacing": 7.0,
+			"FramePadding": [6.0, 3.0],
+			"CellPadding": [10.0, 3.0],
+			"ItemSpacing": [7.0, 6.0]
+		},
+		"FullPalette": [
+			[0.10, 0.10, 0.10, 0.90],
+			[0.40, 0.40, 0.40, 1.00],
+			[0.90, 0.90, 0.90, 0.95],
+			[1.00, 1.00, 1.00, 0.00],
+			[0.95, 0.95, 0.95, 0.85],
+			[0.30, 0.30, 0.30, 0.65],
+			[1.00, 1.00, 1.00, 0.00],
+			[1.00, 1.00, 1.00, 1.00],
+			[0.70, 0.70, 0.70, 0.40],
+			[0.50, 0.50, 0.50, 0.45],
+			[1.00, 1.00, 1.00, 0.83],
+			[1.00, 1.00, 1.00, 0.87],
+			[0.80, 0.80, 0.80, 0.90],
+			[0.98, 0.98, 0.98, 0.90],
+			[0.85, 0.85, 0.85, 0.90],
+			[0.75, 0.75, 0.75, 1.00],
+			[0.58, 0.58, 0.58, 1.00],
+			[0.44, 0.44, 0.44, 1.00],
+			[0.10, 0.10, 0.10, 1.00],
+			[0.30, 0.30, 0.30, 1.00],
+			[0.70, 0.70, 0.70, 1.00],
+			[0.26, 0.59, 0.98, 0.40],
+			[0.26, 0.59, 0.98, 0.67],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.06, 0.53, 0.98, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.50, 0.50, 0.50, 1.00],
+			[0.40, 0.40, 0.40, 1.00],
+			[0.30, 0.30, 0.30, 1.00],
+			[0.10, 0.10, 0.10, 1.00],
+			[0.10, 0.10, 0.10, 0.60],
+			[0.10, 0.10, 0.10, 0.90],
+			[0.26, 0.59, 0.98, 0.31],
+			[0.26, 0.59, 0.98, 0.80],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.85, 0.85, 0.85, 0.97],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.30, 0.40, 0.40, 0.50],
+			[1.00, 1.00, 1.00, 0.00],
+			[0.10, 0.10, 0.10, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.90, 0.70, 0.00, 1.00],
+			[0.26, 0.59, 0.98, 0.40],
+			[0.74, 0.74, 0.74, 1.00],
+			[0.81, 0.81, 0.81, 1.00],
+			[1.00, 1.00, 1.00, 0.00],
+			[0.10, 0.10, 0.10, 0.06],
+			[0.26, 0.59, 0.98, 0.35],
+			[0.80, 0.20, 0.20, 1.00],
+			[0.26, 0.59, 0.98, 1.00],
+			[0.70, 0.70, 0.70, 0.56],
+			[0.80, 0.80, 0.80, 0.35],
+			[0.80, 0.80, 0.80, 0.35]
+		]
 	}
 }
