+++ package/SKSE/Plugins/CommunityShaders/Themes/Mystic.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Mystic Purple",
-	"Description": "Magical purple theme with mystical vibes for spell crafting",
-	"Version": "1.0.0",
+	"Description": "Magical purple theme with mystical vibes perfect for spell crafting and enchanting",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 27.8,
+		"GlobalScale": 0.08,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.55,
 		"Palette": {
 			"Background": [0.2, 0.1, 0.3, 0.9],
 			"Text": [0.95, 0.9, 1.0, 1.0],
 			"Border": [0.6, 0.4, 0.8, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.4, 0.3, 0.5, 1.0],
+			"Error": [1.0, 0.3, 0.6, 1.0],
+			"Warning": [1.0, 0.7, 0.4, 1.0],
+			"RestartNeeded": [0.6, 0.9, 0.7, 1.0],
+			"CurrentHotkey": [0.9, 0.8, 1.0, 1.0],
+			"SuccessColor": [0.7, 0.3, 0.9, 1.0],
+			"InfoColor": [0.8, 0.6, 1.0, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.8, 0.7, 0.95, 1.0],
+			"ColorHovered": [0.9, 0.8, 1.0, 1.0],
+			"MinimizedFactor": 0.6
+		},
+		"Style": {
+			"WindowBorderSize": 2.5,
+			"ChildBorderSize": 1.5,
+			"FrameBorderSize": 2.0,
+			"WindowPadding": [16.0, 14.0],
+			"WindowRounding": 7.0,
+			"IndentSpacing": 10.0,
+			"FramePadding": [9.0, 6.0],
+			"CellPadding": [15.0, 6.0],
+			"ItemSpacing": [10.0, 9.0]
+		},
+		"FullPalette": [
+			[0.95, 0.90, 1.00, 0.90],
+			[0.70, 0.60, 0.85, 1.00],
+			[0.20, 0.10, 0.30, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.18, 0.08, 0.25, 0.85],
+			[0.60, 0.50, 0.75, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.10, 0.05, 0.15, 1.00],
+			[0.35, 0.20, 0.50, 0.40],
+			[0.45, 0.25, 0.60, 0.45],
+			[0.10, 0.05, 0.15, 0.83],
+			[0.15, 0.08, 0.20, 0.87],
+			[0.25, 0.15, 0.40, 0.90],
+			[0.18, 0.10, 0.28, 0.90],
+			[0.30, 0.18, 0.45, 0.90],
+			[0.40, 0.25, 0.55, 1.00],
+			[0.50, 0.35, 0.65, 1.00],
+			[0.60, 0.45, 0.75, 1.00],
+			[0.95, 0.90, 1.00, 1.00],
+			[0.75, 0.65, 0.90, 1.00],
+			[0.35, 0.20, 0.50, 1.00],
+			[0.60, 0.40, 0.80, 0.40],
+			[0.70, 0.50, 0.90, 0.67],
+			[0.80, 0.60, 1.00, 1.00],
+			[0.65, 0.35, 0.95, 1.00],
+			[0.75, 0.45, 1.00, 1.00],
+			[0.85, 0.55, 1.00, 1.00],
+			[0.55, 0.40, 0.70, 1.00],
+			[0.65, 0.50, 0.80, 1.00],
+			[0.80, 0.65, 0.95, 1.00],
+			[0.95, 0.90, 1.00, 1.00],
+			[0.95, 0.90, 1.00, 0.60],
+			[0.95, 0.90, 1.00, 0.90],
+			[0.60, 0.40, 0.80, 0.31],
+			[0.70, 0.50, 0.90, 0.80],
+			[0.80, 0.60, 1.00, 1.00],
+			[0.15, 0.08, 0.22, 0.97],
+			[0.70, 0.45, 0.85, 1.00],
+			[0.60, 0.45, 0.75, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.95, 0.90, 1.00, 1.00],
+			[0.90, 0.70, 1.00, 1.00],
+			[0.90, 0.70, 1.00, 1.00],
+			[0.90, 0.70, 1.00, 1.00],
+			[0.60, 0.40, 0.80, 0.40],
+			[0.30, 0.18, 0.40, 1.00],
+			[0.25, 0.15, 0.35, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.95, 0.90, 1.00, 0.06],
+			[0.60, 0.40, 0.80, 0.35],
+			[0.90, 0.40, 0.60, 1.00],
+			[0.70, 0.50, 0.90, 1.00],
+			[0.40, 0.25, 0.55, 0.56],
+			[0.28, 0.18, 0.40, 0.35],
+			[0.28, 0.18, 0.40, 0.35]
+		]
 	}
 }
