+++ package/SKSE/Plugins/CommunityShaders/Themes/NordicFrost.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Nordic Frost",
-	"Description": "Cool blue-white theme reflecting the harsh Nordic climate",
-	"Version": "1.0.0",
+	"Description": "Cool blue-white theme reflecting the harsh Nordic climate and icy mountain peaks",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 26.0,
+		"GlobalScale": -0.08,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.4,
 		"Palette": {
 			"Background": [0.05, 0.15, 0.25, 0.9],
 			"Text": [0.9, 0.95, 1.0, 1.0],
 			"Border": [0.6, 0.8, 1.0, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.4, 0.45, 0.5, 1.0],
+			"Error": [1.0, 0.3, 0.4, 1.0],
+			"Warning": [1.0, 0.8, 0.3, 1.0],
+			"RestartNeeded": [0.7, 0.9, 0.4, 1.0],
+			"CurrentHotkey": [0.5, 0.8, 1.0, 1.0],
+			"SuccessColor": [0.4, 0.8, 0.6, 1.0],
+			"InfoColor": [0.6, 0.8, 0.9, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.7, 0.85, 0.95, 1.0],
+			"ColorHovered": [0.85, 0.95, 1.0, 1.0],
+			"MinimizedFactor": 0.6
+		},
+		"Style": {
+			"WindowBorderSize": 1.5,
+			"ChildBorderSize": 1.0,
+			"FrameBorderSize": 1.0,
+			"WindowPadding": [12.0, 10.0],
+			"WindowRounding": 8.0,
+			"IndentSpacing": 6.0,
+			"FramePadding": [6.0, 4.0],
+			"CellPadding": [10.0, 4.0],
+			"ItemSpacing": [6.0, 6.0]
+		},
+		"FullPalette": [
+			[0.90, 0.95, 1.00, 0.90],
+			[0.70, 0.80, 0.90, 1.00],
+			[0.05, 0.15, 0.25, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.04, 0.12, 0.20, 0.85],
+			[0.50, 0.65, 0.80, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.03, 0.08, 0.15, 1.00],
+			[0.25, 0.35, 0.50, 0.40],
+			[0.30, 0.40, 0.55, 0.45],
+			[0.03, 0.08, 0.15, 0.83],
+			[0.05, 0.10, 0.18, 0.87],
+			[0.15, 0.25, 0.40, 0.90],
+			[0.08, 0.18, 0.28, 0.90],
+			[0.20, 0.30, 0.45, 0.90],
+			[0.30, 0.40, 0.55, 1.00],
+			[0.40, 0.50, 0.65, 1.00],
+			[0.50, 0.60, 0.75, 1.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.70, 0.80, 0.90, 1.00],
+			[0.25, 0.35, 0.50, 1.00],
+			[0.50, 0.70, 0.90, 0.40],
+			[0.60, 0.75, 0.95, 0.67],
+			[0.70, 0.85, 1.00, 1.00],
+			[0.60, 0.80, 1.00, 1.00],
+			[0.70, 0.85, 1.00, 1.00],
+			[0.80, 0.90, 1.00, 1.00],
+			[0.50, 0.65, 0.80, 1.00],
+			[0.60, 0.75, 0.90, 1.00],
+			[0.80, 0.90, 1.00, 1.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.90, 0.95, 1.00, 0.60],
+			[0.90, 0.95, 1.00, 0.90],
+			[0.50, 0.70, 0.90, 0.31],
+			[0.60, 0.75, 0.95, 0.80],
+			[0.70, 0.85, 1.00, 1.00],
+			[0.04, 0.12, 0.20, 0.97],
+			[0.65, 0.80, 0.95, 1.00],
+			[0.50, 0.65, 0.80, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.50, 0.80, 1.00, 1.00],
+			[0.50, 0.80, 1.00, 1.00],
+			[0.50, 0.80, 1.00, 1.00],
+			[0.50, 0.70, 0.90, 0.40],
+			[0.20, 0.30, 0.45, 1.00],
+			[0.15, 0.25, 0.35, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.95, 1.00, 0.06],
+			[0.50, 0.70, 0.90, 0.35],
+			[0.60, 0.40, 0.80, 1.00],
+			[0.60, 0.75, 0.95, 1.00],
+			[0.30, 0.40, 0.55, 0.56],
+			[0.20, 0.30, 0.45, 0.35],
+			[0.20, 0.30, 0.45, 0.35]
+		]
 	}
 }
