+++ package/SKSE/Plugins/CommunityShaders/Themes/Ocean.json
@@ -1,14 +1,100 @@
 {
 	"DisplayName": "Ocean Blue",
-	"Description": "Cool blue tones inspired by deep ocean waters",
-	"Version": "1.0.0",
+	"Description": "Cool blue tones inspired by deep ocean waters and maritime adventures",
+	"Version": "2.0.0",
 	"Author": "Community Shaders Team",
 	"Theme": {
-		"UseSimplePalette": true,
+		"UseSimplePalette": false,
+		"FontSize": 27.5,
+		"GlobalScale": 0.0,
+		"ShowActionIcons": true,
+		"TooltipHoverDelay": 0.45,
 		"Palette": {
 			"Background": [0.1, 0.2, 0.4, 0.9],
 			"Text": [0.9, 0.95, 1.0, 1.0],
 			"Border": [0.3, 0.5, 0.8, 0.8]
-		}
+		},
+		"StatusPalette": {
+			"Disable": [0.4, 0.45, 0.5, 1.0],
+			"Error": [1.0, 0.4, 0.5, 1.0],
+			"Warning": [1.0, 0.8, 0.3, 1.0],
+			"RestartNeeded": [0.3, 0.8, 0.6, 1.0],
+			"CurrentHotkey": [0.6, 0.9, 1.0, 1.0],
+			"SuccessColor": [0.2, 0.7, 0.9, 1.0],
+			"InfoColor": [0.4, 0.7, 1.0, 1.0]
+		},
+		"FeatureHeading": {
+			"ColorDefault": [0.7, 0.85, 1.0, 1.0],
+			"ColorHovered": [0.8, 0.9, 1.0, 1.0],
+			"MinimizedFactor": 0.65
+		},
+		"Style": {
+			"WindowBorderSize": 2.0,
+			"ChildBorderSize": 1.0,
+			"FrameBorderSize": 1.5,
+			"WindowPadding": [15.0, 13.0],
+			"WindowRounding": 5.0,
+			"IndentSpacing": 8.5,
+			"FramePadding": [8.0, 5.0],
+			"CellPadding": [15.0, 4.0],
+			"ItemSpacing": [9.0, 8.0]
+		},
+		"FullPalette": [
+			[0.90, 0.95, 1.00, 0.90],
+			[0.60, 0.70, 0.85, 1.00],
+			[0.10, 0.20, 0.40, 0.90],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.08, 0.16, 0.32, 0.85],
+			[0.50, 0.60, 0.80, 0.65],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.05, 0.10, 0.20, 1.00],
+			[0.20, 0.30, 0.50, 0.40],
+			[0.30, 0.40, 0.60, 0.45],
+			[0.05, 0.10, 0.20, 0.83],
+			[0.08, 0.15, 0.25, 0.87],
+			[0.15, 0.25, 0.45, 0.90],
+			[0.10, 0.18, 0.35, 0.90],
+			[0.18, 0.28, 0.48, 0.90],
+			[0.25, 0.35, 0.55, 1.00],
+			[0.35, 0.45, 0.65, 1.00],
+			[0.45, 0.55, 0.75, 1.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.70, 0.80, 0.90, 1.00],
+			[0.20, 0.30, 0.50, 1.00],
+			[0.30, 0.50, 0.80, 0.40],
+			[0.40, 0.60, 0.90, 0.67],
+			[0.50, 0.70, 1.00, 1.00],
+			[0.25, 0.55, 0.95, 1.00],
+			[0.35, 0.65, 1.00, 1.00],
+			[0.45, 0.75, 1.00, 1.00],
+			[0.40, 0.50, 0.70, 1.00],
+			[0.50, 0.60, 0.80, 1.00],
+			[0.70, 0.80, 0.95, 1.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.90, 0.95, 1.00, 0.60],
+			[0.90, 0.95, 1.00, 0.90],
+			[0.30, 0.50, 0.80, 0.31],
+			[0.40, 0.60, 0.90, 0.80],
+			[0.50, 0.70, 1.00, 1.00],
+			[0.12, 0.18, 0.30, 0.97],
+			[0.35, 0.55, 0.85, 1.00],
+			[0.50, 0.60, 0.80, 0.50],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.95, 1.00, 1.00],
+			[0.60, 0.80, 1.00, 1.00],
+			[0.60, 0.80, 1.00, 1.00],
+			[0.60, 0.80, 1.00, 1.00],
+			[0.30, 0.50, 0.80, 0.40],
+			[0.20, 0.25, 0.40, 1.00],
+			[0.15, 0.20, 0.35, 1.00],
+			[0.00, 0.00, 0.00, 0.00],
+			[0.90, 0.95, 1.00, 0.06],
+			[0.30, 0.50, 0.80, 0.35],
+			[0.80, 0.40, 0.50, 1.00],
+			[0.40, 0.60, 0.90, 1.00],
+			[0.25, 0.35, 0.55, 0.56],
+			[0.18, 0.25, 0.40, 0.35],
+			[0.18, 0.25, 0.40, 0.35]
+		]
 	}
 }
