+++ src/Menu.cpp
@@ -28,10 +28,9 @@
 #include "Menu/MenuHeaderRenderer.h"
 #include "Menu/OverlayRenderer.h"
 #include "Menu/SettingsTabRenderer.h"
-#include "Menu/ThemeManager.h"
 #include "ShaderCache.h"
 #include "State.h"
-#include "ThemeManager.h"
+#include "Menu/ThemeManager.h"
 #include "TruePBR.h"
 #include "Util.h"
 #include "Util.h"
